+++ src/main/java/com/coderman/api/system/service/impl/RoleServiceImpl.java
@@ -1,11 +1,12 @@
 package com.coderman.api.system.service.impl;
 
-import com.coderman.api.system.converter.RoleConverter;
-import com.coderman.api.system.mapper.MenuMapper;
-import com.coderman.api.system.mapper.RoleMapper;
-import com.coderman.api.system.mapper.RoleMenuMapper;
+import com.coderman.api.common.exception.ServiceException;
 import com.coderman.api.common.pojo.system.Role;
 import com.coderman.api.common.pojo.system.RoleMenu;
+import com.coderman.api.system.converter.RoleConverter;
+import com.coderman.api.system.enums.RoleStatusEnum;
+import com.coderman.api.system.mapper.RoleMapper;
+import com.coderman.api.system.mapper.RoleMenuMapper;
 import com.coderman.api.system.service.RoleService;
 import com.coderman.api.system.vo.PageVO;
 import com.coderman.api.system.vo.RoleVO;
@@ -18,6 +19,7 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.CollectionUtils;
 import tk.mybatis.mapper.entity.Example;
 
+import javax.validation.constraints.NotBlank;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
@@ -35,9 +37,6 @@ public class RoleServiceImpl implements RoleService {
 
     @Autowired
     private RoleMenuMapper roleMenuMapper;
-
-    @Autowired
-    private MenuMapper menuMapper;
 
     /**
      * 角色列表
@@ -66,11 +65,18 @@ public class RoleServiceImpl implements RoleService {
      */
     @Override
     public void add(RoleVO roleVO) {
+        @NotBlank(message = "角色名必填") String roleName = roleVO.getRoleName();
+        Example o = new Example(Role.class);
+        o.createCriteria().andEqualTo("roleName",roleName);
+        int i = roleMapper.selectCountByExample(o);
+        if(i!=0){
+            throw new ServiceException("该角色名已被占用");
+        }
         Role role = new Role();
         BeanUtils.copyProperties(roleVO,role);
         role.setCreateTime(new Date());
         role.setModifiedTime(new Date());
-        role.setStatus(1);
+        role.setStatus(RoleStatusEnum.AVAILABLE.getStatusCode());//默认启用添加的角色
         roleMapper.insert(role);
     }
 
@@ -81,6 +87,10 @@ public class RoleServiceImpl implements RoleService {
     @Transactional
     @Override
     public void deleteById(Long id) {
+        Role role = roleMapper.selectByPrimaryKey(id);
+        if(role==null){
+            throw new ServiceException("要删除的角色不存在");
+        }
         roleMapper.deleteByPrimaryKey(id);
         //删除对应的[角色-菜单]记录
         Example o = new Example(RoleMenu.class);
@@ -88,27 +98,64 @@ public class RoleServiceImpl implements RoleService {
         roleMenuMapper.deleteByExample(o);
     }
 
+    /**
+     * 编辑角色信息
+     * @param id
+     * @return
+     */
     @Override
     public RoleVO edit(Long id) {
         Role role = roleMapper.selectByPrimaryKey(id);
+        if(role==null){
+            throw new ServiceException("编辑的角色不存在");
+        }
         RoleVO roleVO = new RoleVO();
         BeanUtils.copyProperties(role,roleVO);
         return roleVO;
     }
 
+    /**
+     * 更新角色信息
+     * @param id
+     * @param roleVO
+     */
     @Override
     public void update(Long id, RoleVO roleVO) {
+        @NotBlank(message = "角色名必填") String roleName = roleVO.getRoleName();
+        Role dbRole = roleMapper.selectByPrimaryKey(id);
+        if(dbRole==null){
+            throw new ServiceException("要更新的角色不存在");
+        }
+        Example o = new Example(Role.class);
+        o.createCriteria().andEqualTo("roleName",roleName);
+        List<Role> roles = roleMapper.selectByExample(o);
+        if(!CollectionUtils.isEmpty(roles)){
+            Role role = roles.get(0);
+            if(!role.getId().equals(id)){
+                throw new ServiceException("该角色名已被占用");
+            }
+        }
         Role role = new Role();
         BeanUtils.copyProperties(roleVO,role);
         role.setModifiedTime(new Date());
         roleMapper.updateByPrimaryKeySelective(role);
     }
 
+    /**
+     * 角色状态
+     * @param id
+     * @param status
+     */
     @Override
     public void updateStatus(Long id, Boolean status) {
+        Role role = roleMapper.selectByPrimaryKey(id);
+        if(role==null){
+            throw new ServiceException("角色不存在");
+        }
         Role t = new Role();
         t.setId(id);
-        t.setStatus(status?0:1);
+        t.setStatus(status?RoleStatusEnum.DISABLE.getStatusCode():
+                RoleStatusEnum.AVAILABLE.getStatusCode());
         roleMapper.updateByPrimaryKeySelective(t);
     }
